+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Fixed SIGPIPE disposition being set to ignore (#255)
+
 ## [2.13.0] - 2024-03-25
 ### Added
 - Added support for Python 3.12 (#244)